Add shared logging utilities module
